% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lazyfreetext}
\alias{lazyfreetext}
\title{The lazyfreetext function}
\usage{
lazyfreetext(data, freetext, stops = stop_words)
}
\arguments{
\item{data}{name of your dataset (must be tidy)}

\item{freetext}{variable containing your freetext data}

\item{stops}{defaults to 'stop_words', rbind your own to stop_words for added stops}
}
\description{
This function tokenises and lemmatises the data in your tidy dataset
}
\examples{
no examples just yet - need to find some free text data in datasets!
}
