% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lazytf}
\alias{lazytf}
\title{The lazytf function}
\usage{
lazytf(data, word = "lemma", grouping_factor)
}
\arguments{
\item{data}{name of your dataset (data must be tokenised, see lazyfreetext)}

\item{word}{name of your 'token' you are counting, defaults to 'lemma' from lazyfreetext}

\item{grouping_factor}{what group do you want your tf_idf to be created from}
}
\description{
This function quickly provides a term frequency-inverse document frequency
by whatever grouping factors you have in your dataset
}
\examples{
no examples just yet!
}
