% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicktfplotfunction.R
\name{quicktfplot}
\alias{quicktfplot}
\title{The quicktfplot function
This function quickly calculates a term-frequency inverse-document-frequency plot for any grouped data}
\usage{
quicktfplot(data, token = "lemma", grouping_factor, topn)
}
\arguments{
\item{data}{name of your dataset (data must be tokenised, see lazyfreetext)}

\item{token}{name of your 'token' you are counting, defaults to 'lemma' from lazyfreetext}

\item{grouping_factor}{what group do you want your tf_idf to be created from}

\item{topn}{number passed to top_n}
}
\description{
The quicktfplot function
This function quickly calculates a term-frequency inverse-document-frequency plot for any grouped data
}
\examples{
none yet!
}
