% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lazyfreetext}
\alias{lazyfreetext}
\title{The lazyfreetext function}
\usage{
lazyfreetext(data, freetext)
}
\arguments{
\item{data}{name of your dataset (must be tidy)}

\item{freetext}{variable containing your freetext data}
}
\description{
This function tokenises and lemmatises the data in your tidy dataset
}
\examples{
no examples just yet - need to find some free text data in datasets!
}
